refactor(backend: create controller and begin ref): Creates controller from routes, removes views, begins refactor of server and API
Begins the refactoring of the backend to function as API for frontend. Restructuring the single routes folder to routes and controllers for organization and scalability.

Previous architecture for the app was based on MVC, so this transitions dedicated server side code. Files are messy with previous code commented out below for reference, which is temporary as the housekeeping continues. This commit and several moving forward will be messy as sections are polished and structure/architectural decisions evolve.